2.0.0-beta.2: API improvements
* Loggers are functions now. Supports debug-style API
* Logger function defaults to level DEBUG
* Logger function accepts a level name as first parameter
* Added aliases 'verbose' and 'silly' for levels 'log' and 'trace', respectively
* Ulog now supports hooks for logger extension and formatters
* Ulog should now be a drop-in replacement for [debug](https://npmjs.com/package/debug)

New in beta.2: In node version, file descriptor is exposed on console.fd so it can
be used in extensions.

## Backward compatible
v2 Should be a backward-compatible release. But possibly some (very) old browsers are no longer supported
because of the use of Function.prototype.name. Tested on Internet Explorer 11, Edge 41, Opera 49, Chrome 63.
Please test before upgrading and report any issues in the [issue tracker](https://github.com/download/ulog/issues).
Not tested in Safari because I do not have access to a Mac. Help in testing welcome!